+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added collection support to `import-object` command.
 - Added `mark-sensitive` filter action.
 - Automatically remove unused tags from database.
+- Add OpenGraph metadata to profile pages.
 
 ### Changed